Revert "Release/v0.0.7 (#17)"
This reverts commit 6cec795e15acbd8e73509e7ccbc95981ef4ecbb9.
