+++ packages/eslint-config/README.md
@@ -1,0 +1,3 @@
+# `@turbo/eslint-config`
+
+Collection of internal eslint configurations.
