+++ packages/eslint-config/library.js
