+++ packages/twilight-css/package.json
@@ -9,7 +9,7 @@
     "dist"
   ],
   "scripts": {
-    "test": "vitest run",
+    "test": "echo \"Error: no test specified\" && exit 1",
     "build": "tsup"
   },
   "keywords": [
@@ -22,21 +22,15 @@
   "peerDependencies": {
     "tailwindcss": "^3.4.1"
   },
-  "dependencies": {
-    "@types/culori": "^2.1.0",
-    "culori": "^4.0.1"
-  },
   "repository": {
     "type": "git",
     "url": "https://github.com/blazeshomida/twilightcss"
   },
   "devDependencies": {
-    "@vitest/coverage-v8": "^1.4.0",
-    "eslint": "^9.0.0",
+    "@repo/typescript-config": "workspace:*",
+    "@types/culori": "^2.1.0",
+    "culori": "^4.0.1",
     "tsup": "^8.0.2",
-    "tsx": "^4.7.2",
-    "typescript": "^5.3.3",
-    "vite-tsconfig-paths": "^4.3.2",
-    "vitest": "^1.4.0"
+    "typescript": "^5.3.3"
   }
 }
