+++ packages/twilight-css/src/lib/config-definers.ts
@@ -1,0 +1,9 @@
+import { PrimitiveConfig, BaseConfig, Theme } from "../types";
+
+export const definePrimitives = <TConfig extends BaseConfig>(
+  primitives: PrimitiveConfig<TConfig>
+) => primitives;
+
+export const defineTheme = <TConfig extends BaseConfig>(
+  theme: Theme<TConfig>
+) => theme;
