+++ packages/twilight-css/src/lib/createTwilight.ts
@@ -1,18 +1,19 @@
 import plugin from "tailwindcss/plugin";
-import { BaseConfig, PrimitiveConfig, ThemeOptions } from "@/types";
-import { deepMerge } from "@/utils";
+import { BaseConfig, PrimitiveConfig, Theme } from "../types";
 import { handlePrimitives } from "./handlePrimitives";
 import { handleThemes } from "./handleThemes";
+import { deepMerge } from "../utils";
 
 export function createTwilight<TConfig extends BaseConfig>(
   primitives: PrimitiveConfig<TConfig>,
-  themeOptions: ThemeOptions<TConfig>
+  themes: Theme<TConfig>[]
 ) {
   const { twPluginPrimitives, twPresetPrimitives } =
     handlePrimitives(primitives);
-  const { twPresetTokens, twPluginThemes } = handleThemes(themeOptions);
-  const base = deepMerge(twPluginPrimitives, twPluginThemes);
-  const twilightPlugin = plugin(({ addBase }) => addBase(base));
+  const { twPresetTokens, twPluginThemes } = handleThemes(themes);
+  const twilightPlugin = plugin(({ addBase }) => {
+    addBase(deepMerge(twPluginPrimitives, twPluginThemes));
+  });
 
   return {
     twilightPlugin,
