+++ packages/twilight-css/src/lib/handlePrimitives.ts
@@ -1,7 +1,7 @@
-import { BaseConfig, PrimitiveConfig } from "@/types";
-import { objectEntries, objectToCss } from "@/utils";
+import { BaseConfig, PrimitiveConfig } from "../types";
+import { objectEntries, objectToCss } from "../utils";
 import { oklch } from "culori";
-function cleanParam(value: number, precision = 2): number {
+function clean(value: number, precision = 2): number {
   return (
     Math.round(parseFloat((value * 10 ** precision).toFixed(precision))) /
     10 ** precision
@@ -11,30 +11,8 @@ function cleanParam(value: number, precision = 2): number {
 export function handlePrimitives<TConfig extends BaseConfig>(
   primitives: PrimitiveConfig<TConfig>
 ) {
-  const cssProperties: Record<string, string> = {
-    "--clr-inherit": "inherit",
-    "--clr-current": "currentColor",
-    "--clr-transparent": "transparent",
-    "--clr-black": "0% 0 0",
-    "--clr-white": "100% 0 0",
-  };
-  const twPresetPrimitives: Record<string, Record<string, string>> = {
-    inherit: {
-      DEFAULT: "var(--clr-inherit) ",
-    },
-    current: {
-      DEFAULT: "var(--clr-current) ",
-    },
-    transparent: {
-      DEFAULT: "var(--clr-transparent)",
-    },
-    black: {
-      DEFAULT: "oklch(var(--clr-black) / <alpha-value>)",
-    },
-    white: {
-      DEFAULT: "oklch(var(--clr-white) / <alpha-value>)",
-    },
-  };
+  const cssProperties: Record<string, string> = {};
+  const twPresetPrimitives: Record<string, Record<string, string>> = {};
   objectEntries(primitives).forEach(([color, shades]) => {
     if (!shades) return;
     const currentColor: Record<string, string> = {};
@@ -46,8 +24,7 @@ export function handlePrimitives<TConfig extends BaseConfig>(
       const shade = String(shadeProp);
       const colorShade = shade === "DEFAULT" ? color : `${color}-${shade}`;
       const property = `--clr-${colorShade}`;
-      cssProperties[property] =
-        `${Math.round(cleanParam(l) * 100)}% ${cleanParam(c)} ${cleanParam(h)}`;
+      cssProperties[property] = `${clean(l) * 100}% ${clean(c)} ${clean(h)}`;
       currentColor[shade] = `oklch(var(${property}) / <alpha-value>)`;
     });
     twPresetPrimitives[color] = currentColor;
