+++ packages/twilight-css/tsconfig.json
@@ -1,0 +1,6 @@
+{
+  "extends": "@repo/typescript-config/base",
+  "compilerOptions": {
+    "outDir": "dist"
+  }
+}
