--- packages/twilightcss/src/index.ts
@@ -1,7 +1,0 @@
-export {
-  definePrimitives,
-  defineTheme,
-  defineShadcnTheme,
-} from "@/lib/config-definers";
-export { TAILWIND_PRIMITIVES as tailwindPrimitives } from "@/themes/tailwind";
-export { createTwilight } from "@/lib/createTwilight";
