--- packages/twilightcss/vitest.config.mts
@@ -1,11 +1,0 @@
-import { defineConfig } from "vitest/config";
-import tsconfigPaths from "vite-tsconfig-paths";
-
-export default defineConfig({
-  plugins: [tsconfigPaths()],
-  test: {
-    coverage: {
-      reporter: ["text", "html"],
-    },
-  },
-});
