+++ tsconfig.json
@@ -1,0 +1,3 @@
+{
+  "extends": "@repo/typescript-config/base.json"
+}
